# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List the current user's saved prompts
      responses:
        "200":
          description: Prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Save a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptInput"
      responses:
        "201":
          description: Created prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}:
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update a saved prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptUpdate"
      responses:
        "200":
          description: Updated prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deletePrompt
      tags: [prompts]
      summary: Delete a saved prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Prompt:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        body:
          type: string
        category:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - body
        - createdAt
        - updatedAt
    PromptInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        category:
          type: string
      required:
        - title
        - body
    PromptUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        category:
          type: string
...